requirements.txt
# Core
pytz==2020.1
Django==3.0.7
django-configurations==2.2
gunicorn==20.0.4
newrelic==5.14.0.142
requests==2.23.0

# For the persistence stores
psycopg2-binary==2.8.5
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.0.0
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.11.0
Markdown==3.2.2
django-filter==2.2.0

# Developer Tools
ipdb==0.13.2
ipython==7.15.0
mkdocs==1.1.2
flake8==3.8.2

# Testing
mock==4.0.2
factory-boy==2.12.0
django-nose==1.4.6
nose-progressive==1.5.2
coverage==5.1

# Static and Media Storage
django-storages==1.9.1
boto3==1.13.23

